perf(video): native BGR555 u16 framebuffer end-to-end
Both consoles composite in BGR555 already; store it directly instead of
expanding to RGB888 per pixel. Buffers shrink 33% (115.2KB->76.8KB GBA,
69.1KB->46.1KB GBC), removes ~38k (GBA) / ~23k (GBC) per-pixel
conversions + bounds checks per frame, and texture upload shrinks 33%
via SDL_PIXELFORMAT_BGR555 (exact XBGR1555 match for GBA/GBC palettes).

- FFI get_video_buffer -> &[u16]; slice contract documented
- gbc: palette_color() replaces 3 duplicated 5->8 expansion blocks;
  per-scanline slice elides per-pixel bounds checks
- gba: composite stores BGR555 with bit 15 masked; forced blank fill(0x7FFF)
- emulator: generic allocate_aligned<T> replaces u8/i16 twins; splash/mock
  paths pack via single bgr555() helper
- frontend: DUMP_VIDEO expands to RGB888 at dump time to preserve the
  e2e byte-size contract (69120/115200); upload_frame() dedupes 3
  SDL_UpdateTexture sites

+++ core/src/gba/mmu.rs
@@ -193,7 +193,7 @@ impl GbaMmu {
     pub fn tick_system_components(
         &mut self,
         elapsed: u32,
-        video_slice: &mut [u8],
+        video_slice: &mut [u16],
         audio_buf: &mut [i16],
         audio_off: usize,
         speed: f32,
@@ -917,7 +917,7 @@ mod tests {
     // one cycle short of it must raise no interrupt, and the next cycle must.
     #[test]
     fn next_event_is_exact() {
-        let mut video: [u8; 0] = [];
+        let mut video: [u16; 0] = [];
         let mut audio: [i16; 0] = [];
 
         // Timer overflow event. Reload 0xFF00, prescaler 1, IRQ enable:
@@ -957,7 +957,7 @@ mod tests {
         mmu.pending_cycles = 130; // 130/64 = 2 whole ticks owed
         assert_eq!(mmu.read_byte(0x04000100), 2, "derived live counter mid-batch");
 
-        let mut video: [u8; 0] = [];
+        let mut video: [u16; 0] = [];
         let mut audio: [i16; 0] = [];
         mmu.pending_cycles = 0;
         mmu.tick_system_components(130, &mut video, &mut audio, 0, 1.0, &mut ppu, false);
@@ -1127,7 +1127,7 @@ mod tests {
     fn ds_sample_transition_lands_on_overflow() {
         let mut mmu = GbaMmu::new(vec![]);
         let mut ppu = crate::gba::ppu::GbaPpu::new();
-        let mut video: [u8; 0] = [];
+        let mut video: [u16; 0] = [];
         let mut audio: [i16; 0] = [];
 
         // DS A -> both sides, 100% volume, sourced from timer 0 (bit 10 clear).
